Remove duplicate __index from metatable.
__index is supposed to be set when creating the userdata metatable, but
that was getting blown away by a duplicate entry in the function list
for the metatable.  Rather that further complicate things, I decided
that it's probably not a good idea to have table-like semantics
coexisting with object-style functions and turned lump lookups into a
function instead.  I kept the __len metamethod around through, since you
can get the length of many different things in Lua, not just tables.