Begin fleshing out map manipulation functionality.
We do not want to expose the individual objects of the map (Thing,
Linedef, Sidedef, etc.) to Lua.  Instead, we will allow modification to
map objects via getter and setter functions on Map objects.  When
references between objects are required, we will use object ids.

Object ids are assigned upon unmarshalling, however in order to avoid
unexpected behavior while scripting, removing an object will not change
the id of other objects.  I plan for there to be a compact/reindex
function that can be run if the user actually wants to 'fix' the
inevitable holes in the various sequences of object numbers, and I
think that such an operation would be necessary before saving the Map.